requirements-leaderboard-app.txt
scipy~=1.11.3
pandas~=2.1.0

streamlit~=1.27.0
hydra-core~=1.3.2
hydra-colorlog~=1.2.0
omegaconf~=2.3.0
wandb~=0.15.11
gitpython~=3.1.37
